Provide an option to diable linking libunwind
For 32-bit ARM the NDK makes use of two unwinders: libgcc and LLVM's
libunwind. libunwind is needed to provide C++ exception handling
support. libgcc is needed to provide compiler runtime support and as
such its unwinder is also seen by the linker.

These two unwinders are not ABI compatible but do use the same names, so
caution is required to avoid ODR bugs. For 32-bit ARM, the libgcc.a in
the NDK is a linker script that ensures that libunwind is linked before
libgcc, causing the linker to prefer symbols from libunwind to those
from libgcc.

This behavior is the same on buildroot build with gcc arm 8.2

As described in
https://android.googlesource.com/platform/ndk/+/master/docs/BuildSystemMaintainers.md#Unwinding

Change-Id: I3db1ba4613239b71ea1e49e3f429d7cb0dfe26de
Signed-off-by: Cody Xie <cody.xie@rock-chips.com>